Move ship-build military scoring into api.concepts for shared use.
Fleet estimates and scores inference both need the same construction helpers; keeping them under military_score_inference forced a cross-analytic import.
